Try using NAT to access FlightAware
This maybe-archival commit adds support for hooking up our Lambda function to a NAT
from within a VPC with Serverless. I don't like this approach because it
adds an additional $20/month (at least) via NAT Gateway and EC2 data-out
charges, but this _might_ be required depending on how FlightAware is
blacklisting IPs.

That said, I said "maybe-archival" because I was  able to get a FA
flight status page from a quick EC2 instance I made, so this might not
be necessary.