I was mocking wrong
I was trying to mock Selenium so I could load my test pages
"dynamically" upon running my tests. This caused me to maintain a
hard-double of `session.FlightAwareSession` which was going to cause
a lot of heartburn.

I wondered if I was doing it wrong, so I turned to Reddit. As it
happened, [I
was](https://old.reddit.com/r/golang/comments/k326eu/writing_mocks_without_duplicating_my_code_is_this/).

Setting up a local webserver and end-to-end testing against that is much
easier. So that's what this commit does.